feat: WIP support for saving recurrent registers
This commit adds support for saving recurrent registers in the `saveregister.py` file. It introduces a new flag `-r` or `--recurrent` to indicate that the register is recurrent. When the flag is provided, the code creates a `MoneyRecurrentRegister` object and saves it to the `recurrentMoneyDAO`. The recurrence type is set to "monthly" by default, but can be customized using the `-r` or `--recurrent` flag.